rename(branch_traversal): more suitable name
