+++ src/commands/branch-commands/next.ts
@@ -1,5 +1,5 @@
 import yargs from "yargs";
-import { switchBranchAction, TraversalDirection } from "../../actions/stack_traversal";
+import { switchBranchAction, TraversalDirection } from "../../actions/branch_traversal";
 import { execStateConfig } from "../../lib/config";
 import { profile } from "../../lib/telemetry";
